Collision spheres correctly show, making the GameObj classes more minimal with the same functionality

+++ src/gameworld/GameWorld.h
@@ -48,7 +48,7 @@ private:
 	TerrainModel level;
 
 	void CalculateFPS();
-	int elapsedTime;
+	double elapsedTime;
 	int frameCount;
 	int currentTime;
 	int previousTime;
